Process a sorted list of repositories
A sorted list will provide a low-cost way of giving the user a sense of where we are in the processing.  If we are only at the letter `c` then we have a long way to go, but if we are at the letter `t` then then end of processing is near.